handle api order + user service

+++ user-service/src/main/java/vn/com/greencraze/user/dto/response/cart/GetListCartItemResponse.java
@@ -19,7 +19,7 @@ public record GetListCartItemResponse(
         Integer quantity,
         Long variantId,
         String variantName,
-        Long variantQuantity,
+        Integer variantQuantity,
         String sku,
         BigDecimal variantPrice,
         @Nullable
